refactor verficiation and plotting
