+++ emgio/visualization/__init__.py
@@ -1,0 +1,6 @@
+from .static import plot_signals, plot_comparison
+
+__all__ = [
+    'plot_signals',
+    'plot_comparison',
+]
